subcmds fixes
