+++ src/commands/games/steamladder/steamladder.js
@@ -1,4 +1,4 @@
-const { Command, CommandError, SwitchbladeEmbed, EmojiUtils, Constants } = require('../../')
+const { Command, CommandError, SwitchbladeEmbed, EmojiUtils } = require('../../../')
 const countries = require('i18n-iso-countries')
 
 // We're using a Polyfill for Intl, as node doesn't come with all locales for formatting.
@@ -10,18 +10,14 @@ const embedColor = 0x292B2D
 const ladders = ['xp', 'games', 'badges', 'playtime', 'age']
 const regions = ['europe', 'north_america', 'south_america', 'asia', 'africa', 'oceania', 'antarctica']
 
-// TODO: Finish the ladder command
-// TODO: Add profile subcommand
-
 module.exports = class SteamLadder extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'steamladder'
-    this.aliases = ['sl']
-    this.category = 'games'
-    this.subcommands = [new SteamLadderProfile(client, this)]
-    this.parameters = new CommandParameters(this,
-      new StringParameter({
+    super(client, {
+      name: 'steamladder',
+      aliases: ['sl'],
+      category: 'games',
+      parameters: [{
+        type: 'string',
         full: false,
         whitelist: ladders,
         required: false,
@@ -34,8 +30,8 @@ module.exports = class SteamLadder extends Command {
               `**${ladders.map(l => `\`${l}\``).join(', ')}**`
             ].join('\n'))
         }
-      }),
-      new StringParameter({
+      }, {
+        type: 'string',
         required: false,
         whitelist: Object.keys(countries.getNames('en')).map(c => c.toLowerCase()).concat(Object.keys(countries.getNames('en'))).concat(regions),
         missingError: ({ t, prefix }) => {
@@ -49,8 +45,8 @@ module.exports = class SteamLadder extends Command {
               `[${t('commands:steamladder.youCanAlsoUse')}](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2)`
             ].join('\n'))
         }
-      })
-    )
+      }]
+    })
   }
 
   canLoad () {
@@ -113,70 +109,5 @@ module.exports = class SteamLadder extends Command {
     }
 
     return `${position}${flag}${username} - ${typeStrings[ladderType]()}`
-  }
-}
-
-class SteamLadderProfile extends Command {
-  constructor (client, parentCommand) {
-    super(client, parentCommand)
-    this.name = 'profile'
-    this.aliases = ['p']
-    this.parameters = new CommandParameters(this,
-      new StringParameter({
-        full: true,
-        required: true,
-        missingError: 'commands:steamladder.noUser'
-      })
-    )
-  }
-
-  async run ({ t, author, channel, language }, query) {
-    channel.startTyping()
-    const formatter = new Intl.NumberFormat(language)
-    let embed = new SwitchbladeEmbed(author)
-    try {
-      const steamid = await this.client.apis.steam.resolve(query)
-      const {
-        steam_user: user,
-        steam_ladder_info: info,
-        steam_stats: stats,
-        ladder_rank: rank
-      } = await this.client.apis.steamladder.getProfile(steamid)
-      const description = [ EmojiUtils.getFlag(user.steam_country_code) ]
-      if (info.is_staff) description.push(Constants.STEAMLADDER_STAFF)
-      if (info.is_winter_18) description.push(Constants.STEAMLADDER_WINTER)
-      if (info.is_donator) description.push(Constants.STEAMLADDER_DONATOR)
-      if (info.is_top_donator) description.push(Constants.STEAMLADDER_TOP_DONATOR)
-      embed
-        .setAuthor('Steam Ladder', 'https://i.imgur.com/tm9VKhD.png')
-        .setColor(embedColor)
-        .setThumbnail(user.steam_avatar_src)
-        .setTitle(`${user.steam_name} \`${user.steam_id}\``)
-        .setURL(user.steamladder_url)
-        .setDescription(description.join(' '))
-
-        .addField(t('commands:steamladder.level'), [
-          `**${formatter.format(stats.level)}** (${formatter.format(stats.xp)} XP)`,
-          `\`${t('commands:steamladder.inTheWorld', { position: formatter.format(rank.worldwide_xp) })}\``
-        ].join('\n'), true)
-
-        .addField(t('commands:steamladder.playtime'), [
-          t('commands:steamladder.hours', { count: `**${formatter.format(Math.round(stats.games.total_playtime_min / 60))}**` }),
-          `\`${t('commands:steamladder.inTheWorld', { position: formatter.format(rank.worldwide_playtime) })}\``
-        ].join('\n'), true)
-
-        .addField(t('commands:steamladder.games'), [
-          `**${formatter.format(Math.round(stats.games.total_games))}**`,
-          `\`${t('commands:steamladder.inTheWorld', { position: formatter.format(rank.worldwide_games) })}\``
-        ].join('\n'), true)
-
-        .addField(t('commands:steamladder.others'), [
-          t('commands:steamladder.badgesWithCount', { count: `**${formatter.format(Math.round(stats.badges.total))}**` }),
-          t('commands:steamladder.joinedOn', { date: `**${new Intl.DateTimeFormat(language).format(new Date(user.steam_join_date))}**` })
-        ].join('\n'), true)
-    } catch (e) {
-      throw new CommandError(t('commands:steamladder.userNotFound'))
-    }
-    channel.send(embed).then(channel.stopTyping())
   }
 }
